inference-server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
httpx>=0.27
pillow>=10.2
pydantic>=2.6

